Tighten composition admission, MCP hub ownership, and local planning
Split the remaining MCP operations hub by concern, trim unused complete-runtime
fixtures from the MCP conftest, and document the owned smoke / attached edges.

Require COMPOSITION_ADMISSION on complete-runtime composition modules
(AUTHORITY/WIRING/LIFECYCLE/DISCOVERY/REFERENCE/MIXED) with an architecture
ratchet, and report setup weights plus heaviest paths from the runtime
inventory.

Fold infrastructure and high-impact path policy into plan_manifest
[local_planning], and label local-test-ownership.json as exact-selector
overrides only.
