+++ tests/boundary/mcp/conftest.py
@@ -1,19 +1,19 @@
-"""MCP transport boundary test configuration."""
+"""MCP transport boundary test configuration.
+
+Only ``attached_complete_runtime`` is needed by this lane today
+(``test_mcp_inspection_relationship_compaction``). Fresh/authorized complete
+fixtures stay out of the MCP conftest so the lane does not hydrate unused
+portfolio templates.
+"""
 
 from __future__ import annotations
 
 from tests.support.complete_runtime_fixtures import (
     attached_complete_runtime,
-    authorized_complete_runtime,
-    authorized_portfolio_template,
     complete_portfolio_template,
-    fresh_complete_runtime,
 )
 
 __all__ = (
     "attached_complete_runtime",
-    "authorized_complete_runtime",
-    "authorized_portfolio_template",
     "complete_portfolio_template",
-    "fresh_complete_runtime",
 )
