+++ tests/boundary/mcp/test_mcp_entrypoint.py
@@ -1,0 +1,71 @@
+"""MCP process entrypoint surface: help, version, and stdio tool names."""
+
+from __future__ import annotations
+
+import asyncio
+import os
+import subprocess
+import sys
+from importlib.metadata import version
+from pathlib import Path
+
+from jacobian.adapters.mcp.server import create_server
+
+MCP_TOOL_NAMES = {
+    "math.find",
+    "math.run",
+}
+
+def test_mcp_stdio_entrypoint_exposes_stable_math_tools(
+    tmp_path: Path,
+) -> None:
+    async def scenario() -> None:
+        from mcp import Client, StdioServerParameters, stdio_client
+
+        environment = dict(os.environ)
+        environment["JACOBIAN_STATE_DIR"] = str(tmp_path)
+        parameters = StdioServerParameters(
+            command=sys.executable,
+            args=["-m", "jacobian.adapters.mcp.server"],
+            env=environment,
+            cwd=Path.cwd(),
+        )
+        async with Client(
+            stdio_client(parameters),
+            raise_exceptions=True,
+        ) as client:
+            listed = await client.list_tools()
+            assert {tool.name for tool in listed.tools} == MCP_TOOL_NAMES
+
+    asyncio.run(scenario())
+
+
+def test_mcp_entrypoint_has_nonstarting_help() -> None:
+    completed = subprocess.run(
+        [sys.executable, "-m", "jacobian.adapters.mcp.server", "--help"],
+        check=False,
+        capture_output=True,
+        text=True,
+        timeout=20,
+    )
+
+    assert completed.returncode == 0
+    assert "Run the Jacobian MCP server" in completed.stdout
+    assert "--tool-profile" not in completed.stdout
+    assert "--tool-name-profile" not in completed.stdout
+    assert "--reasoning-log-mode" not in completed.stdout
+
+
+def test_mcp_entrypoint_reports_distribution_version() -> None:
+    completed = subprocess.run(
+        [sys.executable, "-m", "jacobian.adapters.mcp.server", "--version"],
+        check=False,
+        capture_output=True,
+        text=True,
+        timeout=20,
+    )
+
+    assert completed.returncode == 0
+    assert completed.stdout.strip() == f"jacobian-mcp {version('jacobian')}"
+
+
