+++ tests/composition/authority/test_verification_workflow.py
@@ -10,6 +10,9 @@ from jacobian.contracts.evidence import WitnessRole
 from jacobian.contracts.results import Conclusion, Verification
 from jacobian.references import ReferenceInstallation
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "AUTHORITY"
+
 
 def _claim(
     reference: ReferenceInstallation,
