+++ tests/composition/portfolio/test_builtin_portfolio_installation.py
@@ -10,6 +10,9 @@ from jacobian.portfolio.result import (
 from jacobian.providers.singular_runtime import singular_provider_runtime
 from jacobian.runtime.model import JacobianRuntime
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "WIRING"
+
 
 def test_builtin_portfolio_installs_cleanly(
     fresh_complete_runtime: JacobianRuntime,
