+++ tests/composition/portfolio/test_plugin_registry_snapshots.py
@@ -27,6 +27,9 @@ from jacobian.plugins.registry import PluginRegistryError
 from jacobian.runtime import create_runtime
 from jacobian.runtime.model import JacobianRuntime
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "WIRING"
+
 
 @pytest.fixture
 def plugin_runtime(
