+++ tests/composition/runtime/test_finite_coverage_authority.py
@@ -5,6 +5,9 @@ from __future__ import annotations
 from jacobian.contracts.capabilities import CapabilityRequest
 from jacobian.contracts.results import ExecutionStatus
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "AUTHORITY"
+
 
 def _request(
     scope: list[str | int],
