+++ tests/composition/runtime/test_graph_coloring_encoding.py
@@ -9,6 +9,9 @@ from jacobian.contracts.capabilities import (
 from jacobian.contracts.results import ExecutionStatus
 from jacobian.runtime.model import JacobianRuntime
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "AUTHORITY"
+
 
 def _encode(runtime: JacobianRuntime) -> CapabilityResult:
     return runtime.core.capabilities.invoke(
