+++ tests/composition/runtime/test_polynomial_collision_verify.py
@@ -7,6 +7,9 @@ from jacobian.contracts.capabilities import (
 )
 from jacobian.contracts.results import ExecutionStatus, InputStatus
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "AUTHORITY"
+
 
 def _rational(value: int) -> dict[str, str]:
     return {"num": str(value), "den": "1"}
