+++ tests/composition/runtime/test_sat_lrat_verification.py
@@ -11,6 +11,9 @@ from jacobian.contracts.capabilities import (
 from jacobian.contracts.results import ExecutionStatus
 from jacobian.runtime.model import JacobianRuntime
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "AUTHORITY"
+
 
 def _verify(runtime: JacobianRuntime, cnf_uri: str, proof: bytes, **extra: object):
     return runtime.core.capabilities.invoke(
