+++ tests/composition/runtime/test_topology_discovery.py
@@ -4,6 +4,9 @@ from __future__ import annotations
 
 from jacobian.contracts.capabilities import CapabilityDiscoveryRequest
 
+# Composition-lane admission category for architecture ratchets.
+COMPOSITION_ADMISSION = "DISCOVERY"
+
 
 def test_homology_intent_discovers_the_domain_owned_operation(
     attached_complete_runtime,
