+++ tests/unit/tooling/test_exact_domain_architecture.py
@@ -69,6 +69,28 @@ def test_lower_tiers_cannot_import_complete_runtime_fixture_bindings(
     )
 
 
+def test_composition_complete_runtime_modules_must_declare_admission(
+    tmp_path: Path,
+) -> None:
+    path = tmp_path / "tests" / "composition" / "runtime" / "test_mixed.py"
+    path.parent.mkdir(parents=True)
+    path.write_text(
+        "def test_case(attached_complete_runtime):\n    assert attached_complete_runtime\n",
+        encoding="utf-8",
+    )
+    (tmp_path / "tests" / "topology.toml").write_text(
+        'version = 1\n[[lanes]]\nname = "composition"\ntier = "composition"\n'
+        'paths = ["tests/composition"]\n',
+        encoding="utf-8",
+    )
+
+    report = check_test_architecture(tmp_path)
+
+    assert any(
+        item.code == "composition-admission-missing" for item in report.violations
+    )
+
+
 def test_non_root_pytest_plugins_are_rejected(tmp_path: Path) -> None:
     conftest = tmp_path / "tests" / "composition" / "conftest.py"
     conftest.parent.mkdir(parents=True)
