+++ tools/check_test_architecture.py
@@ -77,6 +77,26 @@ _EXACT_DOMAIN_INSTALL_ALLOWLIST = frozenset(
     }
 )
 
+_COMPOSITION_ADMISSION_CATEGORIES = frozenset(
+    {
+        "AUTHORITY",
+        "WIRING",
+        "LIFECYCLE",
+        "DISCOVERY",
+        "REFERENCE",
+        "MIXED",
+    }
+)
+_COMPLETE_RUNTIME_FIXTURE_NAMES = frozenset(
+    {
+        "fresh_complete_runtime",
+        "attached_complete_runtime",
+        "authorized_complete_runtime",
+        "complete_portfolio_template",
+        "authorized_portfolio_template",
+    }
+)
+
 
 @dataclass(frozen=True)
 class Violation:
@@ -671,6 +691,71 @@ def _exact_domain_install_violations(tree: ast.AST, relative: str) -> list[Viola
     return violations
 
 
+def _uses_complete_runtime_fixture(tree: ast.AST) -> bool:
+    for node in ast.walk(tree):
+        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)):
+            for arg in (*node.args.args, *node.args.kwonlyargs):
+                if arg.arg in _COMPLETE_RUNTIME_FIXTURE_NAMES:
+                    return True
+        elif isinstance(node, ast.Name) and node.id in _COMPLETE_RUNTIME_FIXTURE_NAMES:
+            return True
+    return False
+
+
+def _composition_admission_value(tree: ast.AST) -> tuple[str | None, int | None]:
+    if not isinstance(tree, ast.Module):
+        return None, None
+    for node in tree.body:
+        if not isinstance(node, ast.Assign) or len(node.targets) != 1:
+            continue
+        target = node.targets[0]
+        if not isinstance(target, ast.Name) or target.id != "COMPOSITION_ADMISSION":
+            continue
+        if isinstance(node.value, ast.Constant) and isinstance(node.value.value, str):
+            return node.value.value, node.lineno
+        return None, node.lineno
+    return None, None
+
+
+def _composition_admission_violations(tree: ast.AST, relative: str) -> list[Violation]:
+    """Composition modules that hydrate a complete runtime must declare admission."""
+
+    path = PurePosixPath(relative)
+    if not _is_under(path, _COMPOSITION_PREFIX):
+        return []
+    if path.name == "conftest.py" or not path.name.startswith("test_"):
+        return []
+    if not _uses_complete_runtime_fixture(tree):
+        return []
+    value, line = _composition_admission_value(tree)
+    if value is None:
+        return [
+            Violation(
+                relative,
+                "composition-admission-missing",
+                (
+                    "declare COMPOSITION_ADMISSION as one of "
+                    f"{sorted(_COMPOSITION_ADMISSION_CATEGORIES)} when using "
+                    "complete-runtime fixtures"
+                ),
+                line or 1,
+            )
+        ]
+    if value not in _COMPOSITION_ADMISSION_CATEGORIES:
+        return [
+            Violation(
+                relative,
+                "composition-admission-invalid",
+                (
+                    f"COMPOSITION_ADMISSION={value!r} is not one of "
+                    f"{sorted(_COMPOSITION_ADMISSION_CATEGORIES)}"
+                ),
+                line,
+            )
+        ]
+    return []
+
+
 def _file_violations(
     file_path: Path,
     project_root: Path,
@@ -719,6 +804,7 @@ def _file_violations(
         violations.extend(_conftest_fixture_violations(tree, relative))
     violations.extend(_non_root_pytest_plugins_violations(tree, relative))
     violations.extend(_exact_domain_install_violations(tree, relative))
+    violations.extend(_composition_admission_violations(tree, relative))
     return violations
